Get user from userID is now working

+++ dat250-jpa-designproto/src/main/java/no/hvl/dat250/gruppe9/DAO/FeedUserDAO.java
@@ -23,7 +23,7 @@ public class FeedUserDAO {
         this.entityManager = entityManagerFactory.createEntityManager();
     }
 
-    public FeedUser getUser(int id){
+    public FeedUser getUser(long id){
         return entityManager.find(FeedUser.class, id);
     }
